chore(refactor): Make parent controllers abstract
